# kabtak/backend/openapi.json
{
  "components": {
    "schemas": {
      "ApplicantProfile": {
        "additionalProperties": false,
        "properties": {
          "annual_family_income_inr": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Family Income Inr"
          },
          "course": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course"
          },
          "domicile_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domicile State"
          },
          "study_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Level"
          },
          "study_year": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Year"
          }
        },
        "title": "ApplicantProfile",
        "type": "object"
      },
...
      "ApplicationType": {
        "enum": [
          "fresh",
          "renewal",
          "unknown"
        ],
        "title": "ApplicationType",
        "type": "string"
      },
      "CheckAccepted": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "poll_after_ms": {
            "default": 2000,
            "title": "Poll After Ms",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "check_id",
          "run_id"
        ],
        "title": "CheckAccepted",
        "type": "object"
      },
      "CheckCreate": {
        "additionalProperties": false,
        "properties": {
          "academic_year": {
            "pattern": "^\\d{4}-\\d{2}$",
            "title": "Academic Year",
            "type": "string"
          },
          "applicant_group": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicant Group"
          },
          "application_type": {
            "$ref": "#/components/schemas/ApplicationType"
          },
          "notice_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Url"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicantProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "programme_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Programme Id",
            "type": "string"
          },
          "save": {
            "default": false,
            "title": "Save",
            "type": "boolean"
          }
        },
        "required": [
          "programme_id",
          "academic_year",
          "application_type"
        ],
        "title": "CheckCreate",
        "type": "object"
      },
      "CheckRead": {
        "properties": {
          "academic_year": {
            "title": "Academic Year",
            "type": "string"
          },
          "applicant_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicant Group"
          },
          "application_type": {
            "$ref": "#/components/schemas/ApplicationType"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_completed_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Completed Run Id"
          },
          "notice_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Url"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicantProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "programme_id": {
            "title": "Programme Id",
            "type": "string"
          },
          "programme_name": {
            "title": "Programme Name",
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunRead"
            },
            "title": "Runs",
            "type": "array"
          },
          "saved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved At"
          }
        },
        "required": [
          "id",
          "programme_id",
          "programme_name",
          "academic_year",
          "application_type",
          "applicant_group",
          "profile",
          "notice_url",
          "saved_at",
          "created_at",
          "expires_at",
          "latest_completed_run_id",
          "runs"
        ],
        "title": "CheckRead",
        "type": "object"
      },
      "CheckUpdate": {
        "additionalProperties": false,
        "properties": {
          "save": {
            "title": "Save",
            "type": "boolean"
          }
        },
        "required": [
          "save"
        ],
        "title": "CheckUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRead": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "reference_time": {
            "format": "date-time",
            "title": "Reference Time",
            "type": "string"
          },
          "report_available": {
            "default": false,
            "title": "Report Available",
            "type": "boolean"
          },
          "retry_of_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of Run Id"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "check_id",
          "kind",
          "status",
          "stage",
          "reference_time",
          "created_at"
        ],
        "title": "RunRead",
        "type": "object"
      },
      "RunStage": {
        "enum": [
          "queued",
          "searching",
          "fetching",
          "extracting",
          "checking",
          "finalizing",
          "finished"
        ],
        "title": "RunStage",
        "type": "string"
      },
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "interrupted"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deadline and evidence checking API.",
    "title": "Kabtak API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/checks": {
      "get": {
        "operationId": "list_checks_v1_checks_get",
        "parameters": [
          {
            "in": "query",
            "name": "saved",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saved"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CheckRead"
                  },
                  "title": "Response List Checks V1 Checks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Checks",
        "tags": [
          "checks"
        ]
      },
      "post": {
        "operationId": "create_check_v1_checks_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Check",
        "tags": [
          "checks"
        ]
      }
    },
    "/v1/checks/{check_id}": {
      "delete": {
        "operationId": "delete_check_v1_checks__check_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Check",
        "tags": [
          "checks"
        ]
      },
      "get": {
        "operationId": "read_check_v1_checks__check_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Check",
        "tags": [
          "checks"
        ]
      },
      "patch": {
        "operationId": "update_check_v1_checks__check_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Check",
        "tags": [
          "checks"
        ]
      }
    },
    "/v1/checks/{check_id}/refresh": {
      "post": {
        "operationId": "refresh_check_v1_checks__check_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Check",
        "tags": [
          "checks"
        ]
      }
    },
...
  }
}